fix: resolve compiler warnings and modernize Windows preset file access
Preserve exclusive creation and add installation/Windows acceptance guidance.

+++ src/dsp/Lfo.h
@@ -20,7 +20,7 @@ public:
  LfoOutput Process(){
   hz_+=slew_*(targetHz_-hz_);
   float raw=0;
-  switch(shape_){case 0:raw=std::sin(6.283185307179586*phase_);break;case 1:raw=1-4*std::abs(phase_-.5);break;case 2:raw=2*phase_-1;break;default:raw=phase_<.5?1:-1;}
+  switch(shape_){case 0:raw=static_cast<float>(std::sin(6.283185307179586*phase_));break;case 1:raw=static_cast<float>(1-4*std::abs(phase_-.5));break;case 2:raw=static_cast<float>(2*phase_-1);break;default:raw=phase_<.5?1.f:-1.f;}
   value_+=edge_*(raw-value_);
   phase_+=hz_/rate_;phase_-=std::floor(phase_);
   for(int i=0;i<4;++i)depths_[i]+=slew_*((i==target_?depth_:0)-depths_[i]);
